post global belajar post API agar global
