+++ db.json
@@ -547,12 +547,6 @@
       "body": "doloremque ex facilis sit sint culpa\nsoluta assumenda eligendi non ut eius\nsequi ducimus vel quasi\nveritatis est dolores"
     },
     {
-      "userId": 10,
-      "id": 100,
-      "title": "Koplak",
-      "body": "Lieur aing"
-    },
-    {
       "userId": 1,
       "id": 1,
       "title": "",
@@ -560,9 +554,21 @@
     },
     {
       "userId": 1,
-      "id": 1615343064107,
-      "title": "Sembarangan bangsat",
-      "body": "sialan Jancuk sue"
+      "id": 1618529024695,
+      "title": "contoh",
+      "body": "contoh 1"
+    },
+    {
+      "userId": 1,
+      "id": 1618529082580,
+      "title": "Biasa",
+      "body": "Biasa"
+    },
+    {
+      "userId": 1,
+      "id": 1618529357134,
+      "title": "Masuk Pak eko",
+      "body": "^_^"
     }
   ]
 }
