+++ src/service/Config.js
@@ -1,0 +1,2 @@
+export const RootPath = 'http://localhost:3004';
+export const OnlineRoot = 'https://jsonplaceholder.typicode.com'
