+++ src/service/Post.js
@@ -1,0 +1,16 @@
+import axios from 'axios';
+import { OnlineRoot, RootPath } from './Config';
+
+const Post = (path, root, data) => {
+    const promise = new Promise((resolve, reject) => {
+        axios.post(`${root ? OnlineRoot : RootPath}/${path}`, data)
+            .then((res) => {
+                resolve(res.data);
+            }, (err) => {
+                reject(err)
+            })
+    })
+    return promise;
+}
+
+export default Post;
